# Conflicts:
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/parser/antlr4/BallerinaParser.java
#	tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/types/xml/XMLLiteralTest.java

+++ bvm/ballerina-core/src/main/java/org/ballerinalang/util/codegen/DefaultValue.java
@@ -30,7 +30,7 @@ public class DefaultValue {
 
     private int valueCPIndex = -1;
     private long intValue;
-    private byte byteValue;
+    private long byteValue;
     private double floatValue;
     private BigDecimal decimalValue;
     private String stringValue;
@@ -97,11 +97,11 @@ public class DefaultValue {
         this.booleanValue = booleanValue;
     }
 
-    public byte getByteValue() {
+    public long getByteValue() {
         return byteValue;
     }
 
-    public void setByteValue(byte byteValue) {
+    public void setByteValue(long byteValue) {
         this.byteValue = byteValue;
     }
 }
